Moved page related routes into separate module and moved common parts into common module
